+++ test/unit/messaging/messaging.spec.ts
@@ -303,6 +303,8 @@ function disableRetries(messaging: Messaging) {
   (messaging as any).messagingRequestHandler.httpClient.retry = null;
 }
 
+class CustomArray extends Array { }
+
 describe('Messaging', () => {
   let mockApp: FirebaseApp;
   let messaging: Messaging;
@@ -597,6 +599,40 @@ describe('Messaging', () => {
       ];
       mockedRequests.push(mockBatchRequest(messageIds));
       return messaging.sendAll([validMessage, validMessage, validMessage])
+        .then((response: BatchResponse) => {
+          expect(response.successCount).to.equal(3);
+          expect(response.failureCount).to.equal(0);
+          response.responses.forEach((resp, idx) => {
+            expect(resp.success).to.be.true;
+            expect(resp.messageId).to.equal(messageIds[idx]);
+            expect(resp.error).to.be.undefined;
+          });
+        });
+    });
+
+    it('should be fulfilled with a BatchResponse given array-like (issue #566)', () => {
+      const messageIds = [
+        'projects/projec_id/messages/1',
+        'projects/projec_id/messages/2',
+        'projects/projec_id/messages/3',
+      ];
+      mockedRequests.push(mockBatchRequest(messageIds));
+      const message = {
+        token: 'a',
+        android: {
+          ttl: 3600,
+        },
+      };
+      const arrayLike = new CustomArray();
+      arrayLike.push(message);
+      arrayLike.push(message);
+      arrayLike.push(message);
+      // Explicitly patch the constructor so that down compiling to ES5 doesn't affect the test.
+      // See https://github.com/firebase/firebase-admin-node/issues/566#issuecomment-501974238
+      // for more context.
+      arrayLike.constructor = CustomArray;
+
+      return messaging.sendAll(arrayLike)
         .then((response: BatchResponse) => {
           expect(response.successCount).to.equal(3);
           expect(response.failureCount).to.equal(0);
@@ -2433,6 +2469,24 @@ describe('Messaging', () => {
           messaging.send({data: arg, topic: 'test'});
         }).to.throw('data must be a non-null object');
       });
+
+      it(`should throw given invalid fcmOptions: ${JSON.stringify(arg)}`, () => {
+        expect(() => {
+          messaging.send({fcmOptions: arg, topic: 'test'});
+        }).to.throw('fcmOptions must be a non-null object');
+      });
+
+      it(`should throw given invalid AndroidFcmOptions: ${JSON.stringify(arg)}`, () => {
+        expect(() => {
+          messaging.send({android: {fcmOptions: arg}, topic: 'test'});
+        }).to.throw('fcmOptions must be a non-null object');
+      });
+
+      it(`should throw given invalid ApnsFcmOptions: ${JSON.stringify(arg)}`, () => {
+        expect(() => {
+          messaging.send({apns: {fcmOptions: arg}, topic: 'test'});
+        }).to.throw('fcmOptions must be a non-null object');
+      });
     });
 
     const invalidDataMessages: any[] = [
@@ -2715,6 +2769,14 @@ describe('Messaging', () => {
         },
       },
       {
+        label: 'Generic fcmOptions message',
+        req: {
+          fcmOptions: {
+            analyticsLabel: 'test.analytics',
+          },
+        },
+      },
+      {
         label: 'Android data message',
         req: {
           android: {
@@ -2816,6 +2878,9 @@ describe('Messaging', () => {
               bodyLocArgs: ['arg1', 'arg2'],
               channelId: 'test.channel',
             },
+            fcmOptions: {
+              analyticsLabel: 'test.analytics',
+            },
           },
         },
         expectedReq: {
@@ -2841,6 +2906,9 @@ describe('Messaging', () => {
               body_loc_key: 'body.loc.key',
               body_loc_args: ['arg1', 'arg2'],
               channel_id: 'test.channel',
+            },
+            fcmOptions: {
+              analyticsLabel: 'test.analytics',
             },
           },
         },
@@ -2963,6 +3031,9 @@ describe('Messaging', () => {
               customKey1: 'custom.value',
               customKey2: {nested: 'value'},
             },
+            fcmOptions: {
+              analyticsLabel: 'test.analytics',
+            },
           },
         },
         expectedReq: {
@@ -2995,6 +3066,9 @@ describe('Messaging', () => {
               },
               customKey1: 'custom.value',
               customKey2: {nested: 'value'},
+            },
+            fcmOptions: {
+              analyticsLabel: 'test.analytics',
             },
           },
         },